backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
reportlab==4.0.9
apscheduler==3.10.4
python-multipart==0.0.6
pypdf==3.17.4
pyjwt==2.8.0
httpx==0.27.0
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
email-validator==2.1.0
sentence-transformers==2.3.1
